pyproject.toml
[tool.poetry]
name = "drf-auto-query"
version = "0.1.0"
description = ""
authors = ["Jakob <jakobverlic66@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_auto_query"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
factory-boy = "^3.2.1"
wheel = "^0.40.0"
setuptools = "^67.8.0"
twine = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
